feat: database refactor with authentication

+++ src/EcoPark.Application/Locations/Delete/DeleteLocationCommandHandler.cs
@@ -8,24 +8,30 @@ public class DeleteLocationCommandHandler(IAggregateRepository<LocationModel> re
         DatabaseOperationResponseViewModel result;
         try
         {
-            await repository.UnitOfWork.StartAsync(cancellationToken);
-
-            var databaseOperationResult = await repository.DeleteAsync(command, cancellationToken);
-
-            if (databaseOperationResult)
+            if (await repository.CheckChangePermissionAsync(command, cancellationToken))
             {
-                await repository.UnitOfWork.SaveEntitiesAsync(cancellationToken);
-                await repository.UnitOfWork.CommitAsync(cancellationToken);
+                await repository.UnitOfWork.StartAsync(cancellationToken);
 
-                result = new DatabaseOperationResponseViewModel("Delete", EOperationStatus.Successful,
-                    "Location was deleted successfully!");
+                var databaseOperationResult = await repository.DeleteAsync(command, cancellationToken);
+
+                if (databaseOperationResult)
+                {
+                    await repository.UnitOfWork.SaveEntitiesAsync(cancellationToken);
+                    await repository.UnitOfWork.CommitAsync(cancellationToken);
+
+                    result = new DatabaseOperationResponseViewModel("Delete", EOperationStatus.Successful,
+                        "Location was deleted successfully!");
+                }
+                else
+                {
+                    await repository.UnitOfWork.RollbackAsync(cancellationToken);
+                    result = new DatabaseOperationResponseViewModel("Delete", EOperationStatus.Failed,
+                        "No Location were found with this id");
+                }
             }
             else
-            {
-                await repository.UnitOfWork.RollbackAsync(cancellationToken);
-                result = new DatabaseOperationResponseViewModel("Delete", EOperationStatus.Failed,
-                    "No Location were found with this id");
-            }
+                result = new DatabaseOperationResponseViewModel("Delete", EOperationStatus.NotAuthorized,
+                                       "You have no permission to delete this location");
                 
         }
         catch (Exception e)
