+++ src/EcoPark.Application/ParkingSpaces/Delete/DeleteParkingSpaceCommandHandler.cs
@@ -9,24 +9,30 @@ public class DeleteParkingSpaceCommandHandler(IAggregateRepository<ParkingSpaceM
 
         try
         {
-            await repository.UnitOfWork.StartAsync(cancellationToken);
-
-            var databaseOperationResult = await repository.DeleteAsync(command, cancellationToken);
-
-            if (databaseOperationResult)
+            if (await repository.CheckChangePermissionAsync(command, cancellationToken))
             {
-                await repository.UnitOfWork.SaveEntitiesAsync(cancellationToken);
-                await repository.UnitOfWork.CommitAsync(cancellationToken);
+                await repository.UnitOfWork.StartAsync(cancellationToken);
 
-                result = new DatabaseOperationResponseViewModel("Delete", EOperationStatus.Successful,
-                    "Parking space was deleted successfully!");
+                var databaseOperationResult = await repository.DeleteAsync(command, cancellationToken);
+
+                if (databaseOperationResult)
+                {
+                    await repository.UnitOfWork.SaveEntitiesAsync(cancellationToken);
+                    await repository.UnitOfWork.CommitAsync(cancellationToken);
+
+                    result = new DatabaseOperationResponseViewModel("Delete", EOperationStatus.Successful,
+                        "Parking space was deleted successfully!");
+                }
+                else
+                {
+                    await repository.UnitOfWork.RollbackAsync(cancellationToken);
+                    result = new DatabaseOperationResponseViewModel("Delete", EOperationStatus.Failed,
+                        "No Parking space were found with this id");
+                }
             }
             else
-            {
-                await repository.UnitOfWork.RollbackAsync(cancellationToken);
-                result = new DatabaseOperationResponseViewModel("Delete", EOperationStatus.Failed,
-                    "No Parking space were found with this id");
-            }
+                result = new DatabaseOperationResponseViewModel("Delete", EOperationStatus.NotAuthorized,
+                                       "You have no permission to delete this parking space");
         }
         catch (Exception e)
         {
